Ensure correct type for both sides of the conditional to prevent compiler warnings

+++ src/internal.h
@@ -158,7 +158,7 @@ T3_WINDOW_LOCAL t3_attr_t _t3_get_attr(int idx);
 T3_WINDOW_LOCAL void _t3_init_attr_map(void);
 T3_WINDOW_LOCAL void _t3_free_attr_map(void);
 
-#define _t3_get_value(s, size) (((s)[0] & 0x80) ? _t3_get_value_int(s, size) : (*(size) = 1, (s)[0]))
+#define _t3_get_value(s, size) (((s)[0] & 0x80) ? _t3_get_value_int(s, size) : (uint32_t) (*(size) = 1, (s)[0]))
 T3_WINDOW_LOCAL uint32_t _t3_get_value_int(const char *s, size_t *size);
 T3_WINDOW_LOCAL size_t _t3_put_value(uint32_t c, char *dst);
 T3_WINDOW_LOCAL int _t3_modifier_hack;
